+++ bindings/python/par_runtime/__init__.py
@@ -16,7 +16,7 @@ from par_runtime.runtime import (
     UsageUpdate,
 )
 
-__version__ = "0.8.5"
+__version__ = "0.8.6"
 
 __all__ = [
     "Runtime",